backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
azure-storage-blob
openai
python-dotenv
pydantic
websockets
langchain
langchain_core
langchain_openai
langchain_google_genai
langgraph
pymupdf4llm
pymupdf
google-genai
Pillow
python-docx
reportlab
docling
